requirements.txt
yt-dl
customtkinter
